move wasm api to its own file to make lib.rs clearer to read
